refactor: change cpu_cores_count type to unsigned and adjust related function signatures

+++ include/core/cpu/index.h
@@ -7,8 +7,7 @@
 #include <math.h>
 #include <time.h>
 void read_cpuinfo_data(FILE *fd, CPUData *data);
-// void show_cpuinfo_data(CPUBlock *cpu_block);
 void show_cpu_data(CPUBlock *cpu_block);
 void read_procstat_cpu_data(FILE *fd, CPUData *data);
-size_t cpu_cores_count(FILE *fd);
+unsigned cpu_cores_count(FILE *fd);
 #endif
